requirements.txt
# --- Core Framework & Orchestration ---
apache-airflow==2.10.2
pendulum>=3.0.0

# --- Airflow Providers (Connectivity & Hooks) ---
# Using specific providers to leverage GSheetsHook, SnowflakeHook, and AwsBaseHook
apache-airflow-providers-amazon>=8.24.0
apache-airflow-providers-google==10.0.0
apache-airflow-providers-snowflake>=5.3.0
apache-airflow-providers-dbt-cloud>=3.9.0

# --- Data Manipulation & Storage ---
# High-performance Parquet and S3 handling
pandas>=2.1.0
pyarrow>=14.0.0
awswrangler>=3.7.2

# --- AWS SDK (Internal Cloud Logic) ---
boto3>=1.34.0
botocore>=1.34.0

# --- Database Connectors (SQLAlchemy Stack) ---
# psycopg2-binary is used for the Postgres Source extraction
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
snowflake-connector-python[pandas]>=3.7.0

# --- Google Services (Sheets & Auth) ---
# Support for GSheetsHook and legacy gspread interactions
gspread>=5.10.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# --- Utilities ---
python-dotenv>=1.0.0